+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.0] - 2026-03-14
+
+### Added
+
+- `SceneDependencies.Block()`: blocks activation of the next dependency group during a load sequence.
+   This allows dependency scenes to prevent their dependents from being loaded until they are fully ready when they are waiting for some multi-frame operation.
+
 ## [2.0.0] - 2026-03-12
 
 ### Changed